Create README.md
This is the repository where I keep the code for my masters thesis which I completed at Columbia University on cognitive diversity and cognitive distance metrics.  It turns out that they can be modified as diversity and distance measures over embeddings of texts and this is where I'm currently going with this.